Trait renforce::agent::StochasticAgent

pub trait StochasticAgent<F: Float, S: Space, A: Space> {
    fn get_action_prop(&self, state: &S::Element, action: &A::Element) -> F;
}

An agent that produces probabilites instead of deterministic actions

Required Methods

fn get_action_prop(&self, state: &S::Element, action: &A::Element) -> F

Returns the probability of this agent performing an action in a given state