Trait renforce::model::DeterministicModel

pub trait DeterministicModel<S: Space, A: Space> {
    fn transition2(&self, curr: &S::Element, action: &A::Element) -> S::Element;
    fn reward2(&self, curr: &S::Element, action: &A::Element) -> f64;
    fn update(&mut self, transition: Transition<S, A>);
}

Represents a deterministic model of an environment When the agent performs a specified action in a specified state, there's only one possible next state

Required Methods

fn transition2(&self, curr: &S::Element, action: &A::Element) -> S::Element

Returns the new state of the agent after performing action in curr

fn reward2(&self, curr: &S::Element, action: &A::Element) -> f64

Returns the reward received when performing action in curr

fn update(&mut self, transition: Transition<S, A>)

Upates the model using information from the given transition