Struct renforce::prelude::PolicyAgent

pub struct PolicyAgent<F: Float, S: Space, A: FiniteSpace, D: DifferentiableFunc<S, A, F>> {
    pub log_func: D,
    // some fields omitted
}

Policy Agent

Explicitly stores a stochastic policy as the softmax of some differentiable function

Fields
log_func: D

The function used by this agent to calculate weights passed into Softmax

Methods

impl<S: Space, A: FiniteSpace, D: DifferentiableFunc<S, A, f64>> PolicyAgent<f64, S, A, D>

fn default(action_space: A, log_func: D) -> PolicyAgent<f64, S, A, D>

Creates a new PolicyAgent with temperature 1.0 used in Softmax

impl<F: Float, S: Space, A: FiniteSpace, D: DifferentiableFunc<S, A, F>> PolicyAgent<F, S, A, D>

fn new(action_space: A, log_func: D, temp: F) -> PolicyAgent<F, S, A, D>

Creates a new PolicyAgent with given parameters

fn temp(self, temp: F) -> PolicyAgent<F, S, A, D>

Updates temp field of self

fn get_temp(&self) -> F

Returns temperature used by agent

fn calc_log_grad(&self, state: &S::Element, action: &A::Element) -> Vec<F>

Calculates the derivative of the log of this function