Module renforce::util

Utilities Module

Modules

approx
Function Approximator Module

chooser
Chooser Module

feature
Feature Extraction Module

graddesc
Gradient Descent Module

table
Table Module

Enums

TimePeriod
Some length of time experienced by an agent

Traits

Chooser
Choose Trait

DifferentiableFunc
A differentiable function taking in (state, action) pairs

Feature
A real-valued feature of elements of some state space

FeatureExtractor
Represents something that extracts features from state-action pairs

GradientDescAlgo
Calculates gradient steps

LogDiffFunc
A function taking in (state, action) pairs whose log can be differentiated

Metric
A type with a notion of distance The distance function should satisfy the triangle inequality (and the other metric properties)

ParameterizedFunc
A function that evaluates its input by making use of some parameters

QFunction
QFunction Trait

VFunction
VFunction Trait