Module renforce::util::approx

Function Approximator Module

Structs

QLinear
Represents multiple linear function approximators, one for each action

VLinear
Represents a linear function approximator f(x) = wT g(x) + b where g: S -> Rn maps states to a vector of features Weights updated using squared error cost C = 1/2(wT g(x) + b - y)2