Struct renforce::util::approx::QLinear

pub struct QLinear<S: Space, A: FiniteSpace> where A::Element: Hash + Eq { /* fields omitted */ }

Represents multiple linear function approximators, one for each action

Methods

impl<S: Space, A: FiniteSpace> QLinear<S, A> where A::Element: Hash + Eq

fn new(action_space: &A) -> QLinear<S, A>

Creates a new, empty QLinear

fn add(&mut self, feat: Box<Feature<S>>)

Adds feat to list of features. Should not be called after any calls to eval or update