Struct renforce::util::approx::VLinear

pub struct VLinear<S: Space> { /* fields omitted */ }

Represents a linear function approximator f(x) = wT g(x) + b where g: S -> Rn maps states to a vector of features Weights updated using squared error cost C = 1/2(wT g(x) + b - y)2

Methods

impl<S: Space> VLinear<S>

fn new() -> VLinear<S>

Creates a new Linear V-Function Approximator

fn with_features(feats: Vec<Box<Feature<S>>>) -> VLinear<S>

Creates a new Linear V-Function Approximator with the given features

fn add_feature(self, feature: Box<Feature<S>>) -> VLinear<S>

Adds the specified feature to the end of the feature vector, giving it a random weight