Trait renforce::util::BinaryFeature

pub trait BinaryFeature<S: Space>: Debug {
    fn b_extract(&self, state: &S::Element) -> bool;
}

A feature of a state pair with a binary value

Required Methods

fn b_extract(&self, state: &S::Element) -> bool

Extracts some binary feature from a given state pair