Trait renforce::trainer::OnlineTrainer

pub trait OnlineTrainer<S: Space, A: Space, T: Agent<S, A>> {
    fn train_step(&mut self, agent: &mut T, transition: Transition<S, A>);
    fn train(&mut self, agent: &mut T, env: &mut Environment<State=S, Action=A>);
}

Represents a way to train an agent online (by interacting with the environment)

Required Methods

fn train_step(&mut self, agent: &mut T, transition: Transition<S, A>)

Performs one training iteration using the given transition

fn train(&mut self, agent: &mut T, env: &mut Environment<State=S, Action=A>)

Automatically trains the agent to perform well in the environment